Change GlBufferData to use multiple data types
The delegate now takes an IntPtr to an array, rather than the array
itself.  The GetPointerToArray method allocates a global pointer to a
copy of the data and provides a callback to deallocate the memory.  Each
new type must be manually handled, but by using pattern matching I can
avoid duplicating code.