Add BaseBindable class
The BaseBindable class prevents code reduplication where we store the
state of the currently bound object (e.g. in ebo or vbo or textures).
The base class implements a static that stores a copy of the currently
bound object, and a check to see if the instance is the currently bound
instance.  It also provides a contract that the inheritors must
complete, to bind.