Add EngineObject classes
RenderObject: Represents an object that will be rendered by the Renderer
class

Transform: Holds a Vector3D Position, a rotation, and other information
to represent the location of a sprite

Camera: Represents the location of a camera's location within the scene.
The main camera is stored in a static variable.

Renderer: Performs the matrix multiplication to get the correct render
position for the RenderObjects.